+++ proton-mail-uniffi/src/mail/user_session.rs
@@ -5,17 +5,18 @@ mod initialization;
 mod labels;
 
 use crate::{
-    core::{datatypes::User, StoredSessionState},
+    core::datatypes::User,
     mail::{MailSessionError, MailSessionResult},
     uniffi_async,
 };
-use futures::TryFutureExt;
-use proton_mail_common::{MailContextError, MailUserContext};
+use proton_mail_common::MailUserContext;
 use stash::stash::Stash;
 use std::sync::Arc;
 
-/// [`MailUserSession`] contains all the relevant information for an active user session, you
-/// obtain one by completing the [`crate::mail::LoginFlow`] or restoring an existing session
+/// [`MailUserSession`] represents an active user session.
+///
+/// This type contains all the relevant information for an active user session.
+/// You obtain one by completing the [`crate::mail::LoginFlow`] or restoring an existing session
 /// with [`crate::mail::MailSession::user_context_from_session`].
 ///
 /// # Initialization
@@ -41,6 +42,18 @@ impl MailUserSession {
 
 #[uniffi::export]
 impl MailUserSession {
+    /// Get the User ID of the current user.
+    #[must_use]
+    pub fn user_id(&self) -> String {
+        self.ctx.user_context().user_id().to_owned().into_inner()
+    }
+
+    /// Get the Session ID of the current user's session.
+    #[must_use]
+    pub fn session_id(&self) -> String {
+        self.ctx.user_context().session_id().to_owned().into_inner()
+    }
+
     /// Log out a session.
     pub async fn logout(&self) -> MailSessionResult<()> {
         let ctx = self.ctx.clone();
@@ -86,71 +99,6 @@ impl MailUserSession {
         uniffi_async(async move {
             let user = ctx.user().await?;
             Ok(user.into())
-        })
-        .await
-    }
-
-    /// Get the state of the session.
-    ///
-    /// If the session has no state (i.e. it was never marked as active), this will return `None`.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database query fails.
-    pub async fn state(&self) -> MailSessionResult<Option<Arc<StoredSessionState>>> {
-        let ctx = self.ctx.clone();
-
-        uniffi_async(async move {
-            let Some(state) = ctx
-                .user_context()
-                .state()
-                .map_err(MailContextError::from)
-                .await?
-            else {
-                return Ok(None);
-            };
-
-            Ok(Some(StoredSessionState::new(state)))
-        })
-        .await
-    }
-
-    /// Mark this session as active.
-    ///
-    /// This updates the last active timestamp of this session in the database.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database operation fails.
-    pub async fn set_active(&self) -> MailSessionResult<()> {
-        let ctx = self.ctx.clone();
-
-        uniffi_async(async move {
-            Ok(ctx
-                .user_context()
-                .set_active()
-                .map_err(MailContextError::from)
-                .await?)
-        })
-        .await
-    }
-
-    /// Return whether the session is active.
-    ///
-    /// A session is considered active if it is the most recent session for the user.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the database operation fails.
-    pub async fn is_active(&self) -> MailSessionResult<bool> {
-        let ctx = self.ctx.clone();
-
-        uniffi_async(async move {
-            Ok(ctx
-                .user_context()
-                .is_active()
-                .map_err(MailContextError::from)
-                .await?)
         })
         .await
     }